fix printf hex float formatting with precision
When formatting a floating point value with %a, we mishandle certain
precisions on architectures where long double is an alias for double. For
example a specifier of  %.12a may emit 13 hex digits after the decimal
instead of 12.

Fix this by tweaking how rounding is performed. Instead of working in
"number of hex digits," work in number of bits to round off the mantissa.
This simplifies the code.

Remove the "extra" round of 8.0, which was unnecessary.